+++ NEWS.md
@@ -1,6 +1,8 @@
-# aqp 1.42 (2022-03-14)
+# aqp 1.42 (2022-03-15)
  * `getArgillicBounds()`, `getCambicBounds()`, `mollic.thickness.requirement()`, `getSurfaceHorizonDepth()` and related functions have been optimized and now work on SoilProfileCollection objects with length > 1
  * Added `reduceSPC()` function for selecting a subset of minimal plus additional specified columns from a SoilProfileCollection
+ * Added new "k-keyword" `.NHZ` for use with SoilProfileCollection `[` method
+   * `object[,,.NHZ]` returns a vector with the number of horizons in each profile
 
 # aqp 1.41 (2022-02-11)
  * `plotSPC()` gains new arguments for controlling horizon depth labels